requirements.txt
numpy
pplpy